requirements.txt
# Data Ingestion
requests>=2.32.0
python-dotenv>=1.0.0

# Spark and Data Processing
pyspark>=3.5.0
pandas>=2.0.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0

# API Server
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Optional: Better ML models
# xgboost>=2.0.0
